requirements.txt
fastapi
uvicorn
streamlit
pandas
numpy
faiss-cpu
sentence-transformers
scikit-learn
requests
beautifulsoup4
groq
python-dotenv